+++ src/actions/cloud-compute/consts.ts
@@ -1,0 +1,5 @@
+import { Duration } from "luxon";
+
+export const DEBUG_MODE_KEEP_TUNNEL_OPEN_DURAION = Duration.fromObject({
+  minutes: 45,
+});
